+++ NEWS.md
@@ -8,8 +8,18 @@
 
 * New PPD module (FIXME: add more details)
 
-* `ppc_scatter()` and `ppc_scatter_avg_grouped()` gain an argument `facet_args` 
-  that many other functions already have.
+* Many functions gain an argument `facet_args` for controlling ggplot2 faceting
+  (many other functions have had this argument for a long time). 
+  The ones that just now got the argument are:
+   - `ppc_error_hist()` 
+   - `ppc_error_hist_grouped()`
+   - `ppc_error_scatter()`
+   - `ppc_scatter()`
+   - `ppc_scatter_avg_grouped()`
+   
+* `ppc_scatter()`, `ppc_scatter_avg()`, and `ppc_scatter_avg_grouped()` gain an
+  argument `ref_line`, which can be set to `FALSE` to turn off the `x=y` line
+  drawn behind the scatterplot.
   
 * All PPC categories now have one or more `_data()` functions that return the 
   data used for plotting (Advances #97). The new ones in this release are:
@@ -17,6 +27,7 @@
    - `ppc_scatter_avg_data()`
    - `ppc_stat_data()` 
    - `ppc_bars_data()`
+   - `ppc_error_data()`
 
 # bayesplot 1.7.0